rename 'services' package to 'jobs'
