+++ jobs/poststop_config.go
@@ -1,4 +1,4 @@
-package services
+package jobs
 
 // NewPostStopConfig ... (temporary, remove by end of v3)
 func NewPostStopConfig(serviceName string, raw interface{}) (*Config, error) {
